requirements.txt
# Core Dependencies
streamlit>=1.28.0
google-generativeai>=0.3.0
pandas>=2.0.0
pymysql>=1.1.0
sqlalchemy>=2.0.0

# Database & Caching
redis>=4.5.0
python-dotenv>=1.0.0

# Security & Validation
cryptography>=41.0.0
bcrypt>=4.0.0
pydantic>=2.0.0

# Performance & Monitoring
asyncio>=3.4.3
aiohttp>=3.8.0
prometheus-client>=0.17.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0
black>=23.7.0
flake8>=6.0.0

# Logging & Utilities
structlog>=23.1.0
click>=8.1.0
rich>=13.0.0
